add daily checkin page
