+++ src/components/AddressManagement.svelte
@@ -1585,12 +1585,6 @@
                 <slot name="overview" />
               {/if}
             {/key}
-
-            {#key selectedWallet || selectedChain}
-              {#if type === "portfolio"}
-                <slot name="dailyCheckin" />
-              {/if}
-            {/key}
           </div>
         </div>
       </div>
