+++ src/entries/newTab/App.svelte
@@ -322,6 +322,20 @@
               </div>
             {/await}
           </Route>
+
+          <Route path="daily-checkin">
+            {#await import("~/layouts/DailyCheckin.svelte")}
+              <div class="flex items-center justify-center h-screen">
+                <Loading />
+              </div>
+            {:then { default: component }}
+              <svelte:component this={component} />
+            {:catch error}
+              <div class="flex items-center justify-center h-screen">
+                Something when wrong! Please reload your browser to try again
+              </div>
+            {/await}
+          </Route>
         </div>
         <div class="fixed bottom-0 left-0 z-30 w-full footer xl:relative">
           <div class="hidden xl:block">
